requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
validators==0.22.0
dnspython==2.4.2
ipaddress==1.0.23
python-whois==0.8.0
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
# Enhanced capabilities
docker==6.1.3
aiohttp==3.9.1
psutil==5.9.6
pefile==2023.2.7
yara-python==4.3.1
# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
